// apps/web/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        DashboardOut: {
            /**
             * Now
             * Format: date-time
             */
            now: string;
            next_performance: components["schemas"]["PerformanceOut"] | null;
            /** Next 7 Days */
            next_7_days: number;
            /** Next 30 Days */
            next_30_days: number;
            /** Upcoming Sessions */
            upcoming_sessions: components["schemas"]["PerformanceOut"][];
            /** Published Events */
            published_events: number;
            /** Draft Events */
            draft_events: number;
            /** Draft Stories */
            draft_stories: number;
            /** Published Stories */
            published_stories: number;
            /** Video Count */
            video_count: number;
            /** Videos Without Id */
            videos_without_id: number;
            /** Sessions Without Tickets */
            sessions_without_tickets: number;
            /** Recent Media */
            recent_media: components["schemas"]["MediaAssetOut"][];
            /** Recent Edits */
            recent_edits: components["schemas"]["AuditOut"][];
            /** Unresolved Messages */
            unresolved_messages: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}